Added: cross entropy loss calculation

+++ src/kernel/math.cu
@@ -799,7 +799,7 @@ __global__ void QKVMatmulKernel(
 // we will fuse everything here.
 __global__ void oneHotKernel(
     int *y, // (B, T )
-    int *out, // (B*T, vocab_size)
+    float *out, // (B*T, vocab_size)
     int N, // number of elements in y as if it was flat which it is.
     int width)
 {
@@ -820,7 +820,7 @@ __global__ void oneHotKernel(
 // SO: for our each B,T we will have a unique vocab as proballity.
 __global__ void crossEntropyLoss(
     float *x, // (B, T, vocab_size)
-    int *y, // (B, T)
+    float *y, // (vocab_size)
     float *out, // N loss across all B,T basically.
     int seq_len,
     int vocab_size,
@@ -830,12 +830,19 @@ __global__ void crossEntropyLoss(
     // The way we define the perfectness is 1 here.
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
 
+    // so here we basically want to read from the index of x to y and compute
+    // the proballty 
+
     if (idx < N)
     {
-        int label = y[idx];
+        int label = y[idx]; // lookup psoition for x
         float p = x[idx * vocab_size + label]; // read from this x that vocab_size dimension 
         out[idx] = -logf(fmaxf(p, 1e-12f));
     }
+
+    // cross entropy is calcuated only for 1 position in the actual label
+    // meaning we will have [0,0,0,...1,0,0] somewhere
+    // for each B,T shape you have 1 thats B,T
 }
 
 extern "C"
@@ -844,7 +851,7 @@ extern "C"
     void CrossEntropy(
         float *x,
         int *y,
-        int *oneHotOut,
+        float *oneHotOut,
         float *lossOut,
         int batch_size,
         int seq_len,
